+++ src/MapFromJson.php
@@ -11,6 +11,7 @@ class MapFromJson
 {
     public const INT_TO_DATETIME = 'intToDateTime';
     public const STRING_TO_ACTION = 'stringToAction';
+    public const STRING_TO_ID = 'stringToId';
 
 
     public function __construct(